EXCEPT_RANDOM_MANUALLY_EXIT
It provides a simple interactive application that allows the user to select between generating a random number or entering a number manually. The program then allows the user to divide the chosen number by 0, and it catches the ZeroDivisionError and prints an appropriate message in case a user tries to divide by 0.